[Bronze V] Title: 이상한 기호, Time: 104 ms, Memory: 14168 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -3,16 +3,16 @@ import java.lang.*;
 import java.io.*;
 
 class Main {
-    static int A, B;
+    static long A, B;
     
 	public static void main (String[] args) throws java.lang.Exception {
 	    BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 	    StringTokenizer st = new StringTokenizer(br.readLine());
 	    
-	    A = Integer.parseInt(st.nextToken());
-	    B = Integer.parseInt(st.nextToken());
+	    A = Long.parseLong(st.nextToken());
+	    B = Long.parseLong(st.nextToken());
 	  
-	    int result = A*A  - B*B;
+	    long result = A*A  - B*B;
 	    System.out.println(result);
 	}
 }
